simulator/gateway/requirements.txt
flask==3.0.3
requests==2.32.3
bsdiff4==1.2.4
python-can==4.3.1
paho-mqtt
cryptography
grpcio
grpcio-tools
protobuf
